Fix undeploy hanging issue
